Delete Confirmation Dialog

+++ EmployeeManagementWeb/Pages/DisplayEmployeeBase.cs
@@ -1,4 +1,5 @@
-﻿using EmployeeManagementModels.Models;
+﻿using ComponentsLibrary;
+using EmployeeManagementModels.Models;
 using EmployeeManagementWeb.Services;
 using Microsoft.AspNetCore.Components;
 using System;
@@ -27,12 +28,23 @@ namespace EmployeeManagementWeb.Pages
         [Inject]
         public NavigationManager NavigationManager { get; set; }
 
-        protected async Task Delete_Click()
+        public ConfirmBase DeleteConfirmation { get; set; }
+
+        protected void Delete_Click()
         {
-            await EmployeeService.DeleteEmployee(Employee.EmployeeId);
-            await OnEmployeeDelete.InvokeAsync(Employee.EmployeeId);
-            //NavigationManager.NavigateTo("/", true);
+            DeleteConfirmation.Show();
+          
         }
+        protected async Task ConfirmDelete_Click(bool deleteConfirmed)
+        {
+            if (deleteConfirmed)
+            {
+                await EmployeeService.DeleteEmployee(Employee.EmployeeId);
+                await OnEmployeeDelete.InvokeAsync(Employee.EmployeeId);
+            }
+        }
+
+       
         protected async Task CheckBoxChanged(ChangeEventArgs e)
         {
             await OnEmployeeSelection.InvokeAsync((bool)e.Value);
